third_party/blink/web_tests/http/tests/loading/bad-server-subframe.html

<script>
if (window.testRunner)
    testRunner.dumpAsText();

window.onload = function() {
  // Navigate the iframe from onload, so that it happens after all loading
  // logs from the main frame are complete. This avoids log ordering races
  // if the iframe is loaded in another process due to site isolation.
  var i = document.createElement("iframe");
  i.name = "f1";
  i.src = "http://localhost:54321";
  document.body.appendChild(i);
}
</script>
This is a test of load callbacks. It is only useful inside the regression test tool.<br>
It also assumes there is no web server running locally on port 54321.<br>
